mc_wiki/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minecraft-wiki-local"
version = "0.1.0"
description = "Offline MediaWiki crawler and structured indexer for Minecraft Wiki"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27,<1",
  "mwparserfromhell>=0.6.6,<1",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
